+++ src/components/Home/2.Achievement.tsx
@@ -11,7 +11,7 @@ const Page = (props: Props) => {
         <div className="container">
           <Title>
             <h1>성장하는 언틸드</h1>
-            <h3>팀원과 함께 끊임없이 성장해 나가고 있어요.</h3>
+            <h3>팀원들과 함께 끊임없이 성장해 나가고 있어요.</h3>
           </Title>
           <BoxList>
             <Box></Box>
